requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.2.0
yfinance>=0.2.36
numpy>=1.26.0
scipy>=1.11.0
matplotlib>=3.8.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
